+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - [[#117]](https://github.com/Marcin-Migdal/m-component-library/issues/117) **[PATCH]** Implement improved version of the gap prop in Row component, that enabled possibility of configuring gap css property for all breakpoints
 - [[#120]](https://github.com/Marcin-Migdal/m-component-library/issues/120) **[PATCH]** Fix button and colorPicker components types, created SimpleInputProps for inputs without floating label
 - [[#119]](https://github.com/Marcin-Migdal/m-component-library/issues/119) **[PATCH]** Fix Breadcrumb background
+- [[#110]](https://github.com/Marcin-Migdal/m-component-library/issues/110) **[PATCH]** Fix dropdown menu styles
 
 ### Release date 25.10.2024 - Release 1.0.6